Add auto: false to livereload calls so build works
Build worked when env was production but jammed if it was not because of
the livereload server being up and running. That can be prevented by
using option auto: false, but this introduces an issue where the
livereload server is not started at all if auto: false call gets executed
first (and it would if the assets were built first). Thus, start the
watch task first.